Display username in header.
