+++ web-gui/src/app/page-components/classroom/classroom.component.html
@@ -1,17 +1,5 @@
 <div class="container">
-  <app-menu-bar>
-    <span class="spacer"></span>
-
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() " (click)="createTicket()" matTooltip="Ticket erstellen"><mat-icon >live_help</mat-icon></button>
-    <!--<button mat-icon-button color="primary" *ngIf="isAuthorized() && classroomService.isJoined()" (click)="createConference()" matTooltip="Konferenzeinstellungen"><mat-icon >duo</mat-icon></button>-->
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && !conferenceService.conferenceWindowOpen" (click)="this.conferenceService.createConference()" matTooltip="Konferenz öffnen"><mat-icon >videocam</mat-icon></button>
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && conferenceService.conferenceWindowOpen" (click)="this.conferenceService.closeConference()" matTooltip="Konferenz schliessen"><mat-icon >videocam_off</mat-icon></button>
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && !isInClassroom(currentUser.userId)" (click)="this.classroomService.showUser()" matTooltip="Nutzer sichtbar machen"><mat-icon >visibility</mat-icon></button>
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && isInClassroom(currentUser.userId)" (click)="this.classroomService.hideUser()" matTooltip="Nutzer unsichtbar machen"><mat-icon >visibility_off</mat-icon></button>
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && conferenceService.conferenceWindowOpen && !isInConferenceId(currentUser.userId)" (click)="this.classroomService.showConference()" matTooltip="Konferenz sichtbar machen"><mat-icon >lock_open</mat-icon></button>
-    <button mat-icon-button color="primary" *ngIf="classroomService.isJoined() && conferenceService.conferenceWindowOpen && isInConferenceId(currentUser.userId)" (click)="this.classroomService.hideConference()" matTooltip="Konferenz unsichtbar machen"><mat-icon >lock</mat-icon></button>
-    <button mat-icon-button class="warn" *ngIf="classroomService.isJoined()" (click)="leaveClassroom()" matTooltip="Online Übungsraum verlassen"><mat-icon >cancel</mat-icon></button>
-  </app-menu-bar>
+  <app-menu-bar></app-menu-bar>
 </div>
 <div class="ticketOverview">
   <div class="ticketColumn users">
@@ -22,7 +10,7 @@
         <span class="meta">
           <mat-icon *ngIf="this.classroomService.isCurrentUserAuthorized() && classroomService.isNotSelf(user)"
                     class="material-icons call" color="accent" (click)="inviteToConference([user])">call</mat-icon>
-          <mat-icon *ngIf="isInConference(user)" class="material-icons call" color="accent" (click)="this.classroomService.joinConference(user)">phone_callback</mat-icon>
+          <mat-icon *ngIf="classroomService.isInConference(user)" class="material-icons call" color="accent" (click)="this.classroomService.joinConference(user)">phone_callback</mat-icon>
           <!-- <mat-icon *ngIf="classroomService.isInConference(user) && !classroomService.hasOpenConference(user)" class="material-icons call" color="accent" (click)="joinConference(classroomService.isInConference(user).creator)">meeting_room</mat-icon> -->
         </span>
       </mat-card>
