+++ web-gui/src/app/page-components/menu-bar/menu-bar.component.scss
@@ -2,11 +2,13 @@
   background-color: transparent;
   box-sizing: border-box;
   height: 48px;
-  padding: 0;
   display: flex;
   align-items: center;
-  padding-right: 8px;
-  padding-left: 8px;
+  padding: 0 8px;
   box-shadow: none;
   border: 1px solid rgba(0, 0, 0, 0.12);
+}
+
+.spacer {
+  flex: 1 1 auto;
 }
