+++ web-gui/src/app/page-components/menu-bar/menu-bar.component.ts
@@ -1,4 +1,6 @@
 import { Component } from '@angular/core';
+import {ClassroomService} from "../../service/classroom.service";
+import {ConferenceService} from "../../service/conference.service";
 
 @Component({
   selector: 'app-menu-bar',
@@ -6,4 +8,11 @@ import { Component } from '@angular/core';
   styleUrls: ['./menu-bar.component.scss']
 })
 export class MenuBarComponent {
+
+  constructor(
+    public classroomService: ClassroomService,
+    public conferenceService: ConferenceService
+  ) {
+  }
+
 }
